move UI components to components/ directory
