+++ src/components/InfoBox.scss
@@ -1,5 +1,5 @@
-@import "shared";
-@import "sharingButtons";
+@import "../shared";
+@import "./shareButtons";
 
 $max-height: 500px;
 $overlay-height: auto;
