+++ src/components/InfoButton.scss
@@ -1,4 +1,4 @@
-@import "shared";
+@import "../shared";
 
 .info-button {
