+++ src/components/Intro.scss
@@ -1,4 +1,4 @@
-@import "shared";
+@import "../shared";
 
 .intro {
   h2 {
