+++ src/components/Menu.scss
@@ -1,4 +1,4 @@
-@import "shared";
+@import "../shared";
 
 $medium_breakpoint: 1250px;
 $mobile_breakpoint: 1120px;
